Add user_signed_in to the code_module controller

+++ app/controllers/code_modules_controller.rb
@@ -2,40 +2,65 @@ class CodeModulesController < ApplicationController
 
     # Method for all CodeModules
     def index
-       codeModules = CodeModule.all
-        render json: codeModules 
+        # Making sure only signed in users can access
+        if user_signed_in?
+            codeModules = CodeModule.all
+            render json: codeModules 
+        else
+            render status: 403, plain: "Please Sign In to Access"
+        end
     end
     # Method for making a new CodeModule
     def create
-        codeModule = CodeModule.create(codeModule_params)
-        if codeModule.valid?
-            render json: codeModule
+        # Making sure only signed in users can access
+        if user_signed_in?
+            codeModule = CodeModule.create(codeModule_params)
+            if codeModule.valid?
+                render json: codeModule
+            else
+                render json: codeModule.errors, status: :unprocessable_entity
+            end
         else
-            render json: codeModule.errors, status: :unprocessable_entity
+            render status: 403, plain: 'Please Sign In to Access'
         end
     end
-    # Method to show just 1 
+    # Method to show just 1 CodeModule
     def show
-        codeModule = CodeModule.find(params[:id])
-        render json: codeModule 
+        # Making sure only signed in users can access
+        if user_signed_in?
+            codeModule = CodeModule.find(params[:id])
+            render json: codeModule 
+        else
+            render status: 403, plain: 'Please Sign In to Access'
+        end
     end
     # Method to update a CodeModule
     def update
-        codeModule = CodeModule.find(params[:id])
-        codeModule.update(codeModule_params)
-        if codeModule.valid?
-            render json: codeModule
+        # Making sure only signed in users can access
+        if user_signed_in?
+            codeModule = CodeModule.find(params[:id])
+            codeModule.update(codeModule_params)
+            if codeModule.valid?
+                render json: codeModule
+            else
+                render json: codeModule.errors
+            end
         else
-            render json: codeModule.errors
+            render status: 403, plain: 'Please Sign In to Access'
         end
     end
     # Method to delete a CodeModule
     def destroy
-        codeModule = CodeModule.find(params[:id])
-        if codeModule.destroy
-            render json: codeModule
+        # Making sure only signed in users can access
+        if user_signed_in?
+            codeModule = CodeModule.find(params[:id])
+            if codeModule.destroy
+                render json: codeModule
+            else
+                render json: codeModule.errors
+            end
         else
-            render json: codeModule.errors
+            render status: 403, plain: 'Please Sign In to Access'
         end
     end
     # Strong params below
